ftime to gettimeofday conversion, time adjustment bug fix for log rotation, latency is now a float
git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@309 5f96b256-904b-4d8d-8c98-d829582c6739

+++ cgi/tac.c
@@ -3,7 +3,7 @@
  * TAC.C - Nagios Tactical Monitoring Overview CGI
  *
  * Copyright (c) 2001-2003 Ethan Galstad (nagios@nagios.org)
- * Last Modified: 02-25-2003
+ * Last Modified: 05-08-2003
  *
  * This CGI program will display the contents of the Nagios
  * log file.
@@ -123,13 +123,13 @@ double min_host_execution_time=-1.0;
 double max_host_execution_time=-1.0;
 double total_host_execution_time=0.0;
 double average_host_execution_time=-1.0;
-int min_service_latency=-1;
-int max_service_latency=-1;
-unsigned long total_service_latency=0L;
+double min_service_latency=-1.0;
+double max_service_latency=-1.0;
+double total_service_latency=0.0;
 double average_service_latency=-1.0;
-int min_host_latency=-1;
-int max_host_latency=-1;
-unsigned long total_host_latency=0L;
+double min_host_latency=-1.0;
+double max_host_latency=-1.0;
+double total_host_latency=0.0;
 double average_host_latency=-1.0;
 
 int flapping_services=0;
@@ -543,9 +543,9 @@ void analyze_status_data(void){
 
 			total_active_service_checks++;
 
-			if(min_service_latency==-1 || temp_servicestatus->latency<min_service_latency)
+			if(min_service_latency==-1.0 || temp_servicestatus->latency<min_service_latency)
 				min_service_latency=temp_servicestatus->latency;
-			if(max_service_latency==-1 || temp_servicestatus->latency>max_service_latency)
+			if(max_service_latency==-1.0 || temp_servicestatus->latency>max_service_latency)
 				max_service_latency=temp_servicestatus->latency;
 
 			if(min_service_execution_time==-1.0 || temp_servicestatus->execution_time<min_service_execution_time)
@@ -664,9 +664,9 @@ void analyze_status_data(void){
 
 			total_active_host_checks++;
 
-			if(min_host_latency==-1 || temp_hoststatus->latency<min_host_latency)
+			if(min_host_latency==-1.0 || temp_hoststatus->latency<min_host_latency)
 				min_host_latency=temp_hoststatus->latency;
-			if(max_host_latency==-1 || temp_hoststatus->latency>max_host_latency)
+			if(max_host_latency==-1.0 || temp_hoststatus->latency>max_host_latency)
 				max_host_latency=temp_hoststatus->latency;
 
 			if(min_host_execution_time==-1.0 || temp_hoststatus->execution_time<min_host_execution_time)
@@ -925,7 +925,7 @@ void display_tac_overview(void){
 	printf("</tr>\n");
 	printf("<tr>\n");
 	printf("<td align=left valign=center class='perfItem'><a href='%s?type=%d' class='perfItem'>Service Check Latency:</a></td>",EXTINFO_CGI,DISPLAY_PERFORMANCE);
-	printf("<td valign=top class='perfValue' nowrap><a href='%s?type=%d' class='perfValue'>%d / %d / %2.3f sec</a></td>\n",EXTINFO_CGI,DISPLAY_PERFORMANCE,min_service_latency,max_service_latency,average_service_latency);
+	printf("<td valign=top class='perfValue' nowrap><a href='%s?type=%d' class='perfValue'>%.2f / %.2f / %.3f sec</a></td>\n",EXTINFO_CGI,DISPLAY_PERFORMANCE,min_service_latency,max_service_latency,average_service_latency);
 	printf("</tr>\n");
 	printf("<tr>\n");
 	printf("<td align=left valign=center class='perfItem'><a href='%s?type=%d' class='perfItem'>Host Check Execution Time:</a></td>",EXTINFO_CGI,DISPLAY_PERFORMANCE);
@@ -933,7 +933,7 @@ void display_tac_overview(void){
 	printf("</tr>\n");
 	printf("<tr>\n");
 	printf("<td align=left valign=center class='perfItem'><a href='%s?type=%d' class='perfItem'>Host Check Latency:</a></td>",EXTINFO_CGI,DISPLAY_PERFORMANCE);
-	printf("<td valign=top class='perfValue' nowrap><a href='%s?type=%d' class='perfValue'>%d / %d / %2.3f sec</a></td>\n",EXTINFO_CGI,DISPLAY_PERFORMANCE,min_host_latency,max_host_latency,average_host_latency);
+	printf("<td valign=top class='perfValue' nowrap><a href='%s?type=%d' class='perfValue'>%.2f / %.2f / %2.3f sec</a></td>\n",EXTINFO_CGI,DISPLAY_PERFORMANCE,min_host_latency,max_host_latency,average_host_latency);
 	printf("</tr>\n");
 	printf("<tr>\n");
 	printf("<td align=left valign=center class='perfItem'><a href='%s?host=all&serviceprops=%d' class='perfItem'># Active Host / Service Checks:</a></td>",STATUS_CGI,SERVICE_ACTIVE_CHECK);
@@ -1217,8 +1217,10 @@ void display_tac_overview(void){
 	printf("</tr>\n");
 	printf("</table>\n");
 
+	/*
 	printf("</tr>\n");
 	printf("</table>\n");
+	*/
 
 	printf("</p>\n");
